# Conflicts:
#	linkis-commons/linkis-common/src/test/java/org/apache/linkis/common/utils/ByteTimeUtilsTest.java

+++ linkis-public-enhancements/linkis-basedata-manager/src/main/java/org/apache/linkis/basedatamanager/server/restful/RmExternalResourceProviderRestfulApi.java
@@ -112,8 +112,13 @@ public class RmExternalResourceProviderRestfulApi {
       httpMethod = "DELETE")
   @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
   public Message remove(HttpServletRequest request, @PathVariable("id") Long id) {
-    ModuleUserUtils.getOperationUser(
-        request, "Remove a Resource manager External Resource Provider Record,id:" + id.toString());
+    String username =
+        ModuleUserUtils.getOperationUser(
+            request,
+            "Try to remove resource external resource provider record with id:" + id.toString());
+    if (!Configuration.isAdmin(username)) {
+      return Message.error("User '" + username + "' is not admin user[非管理员用户]");
+    }
     boolean result = rmExternalResourceProviderService.removeById(id);
     return Message.ok("").data("result", result);
   }
