+++ linkis-public-enhancements/linkis-basedata-manager/src/main/java/org/apache/linkis/basedatamanager/server/restful/UdfManagerRestfulApi.java
@@ -32,6 +32,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 
+import java.util.Date;
+
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.Api;
@@ -89,6 +91,8 @@ public class UdfManagerRestfulApi {
         new QueryWrapper<>(udfManagerEntity).eq("user_name", udfManagerEntity.getUserName());
     UdfManagerEntity udfManager = udfManagerService.getOne(queryWrapper);
     if (udfManager == null) {
+      udfManagerEntity.setCreateTime(new Date());
+      udfManagerEntity.setUpdateTime(new Date());
       boolean result = udfManagerService.save(udfManagerEntity);
       return Message.ok("").data("result", result);
     } else {
@@ -103,7 +107,12 @@ public class UdfManagerRestfulApi {
       httpMethod = "DELETE")
   @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
   public Message remove(HttpServletRequest request, @PathVariable("id") Long id) {
-    ModuleUserUtils.getOperationUser(request, "Remove a UDF Manager Record,id:" + id.toString());
+    String username =
+        ModuleUserUtils.getOperationUser(
+            request, "Remove a UDF Manager Record,id:" + id.toString());
+    if (!Configuration.isAdmin(username)) {
+      return Message.error("User '" + username + "' is not admin user[非管理员用户]");
+    }
     boolean result = udfManagerService.removeById(id);
     return Message.ok("").data("result", result);
   }
@@ -121,7 +130,15 @@ public class UdfManagerRestfulApi {
     if (!Configuration.isAdmin(username)) {
       return Message.error("User '" + username + "' is not admin user[非管理员用户]");
     }
-    boolean result = udfManagerService.updateById(udfManagerEntity);
-    return Message.ok("").data("result", result);
+    QueryWrapper<UdfManagerEntity> queryWrapper =
+        new QueryWrapper<>(udfManagerEntity).eq("user_name", udfManagerEntity.getUserName());
+    UdfManagerEntity udfManager = udfManagerService.getOne(queryWrapper);
+    if (udfManager == null) {
+      udfManager.setUpdateTime(new Date());
+      boolean result = udfManagerService.updateById(udfManagerEntity);
+      return Message.ok("").data("result", result);
+    } else {
+      return Message.error("The username already exists,Please update again!");
+    }
   }
 }
